Updated Socket Error Handling
Updated poorly designed socket error handling with new logic for improved accuracy and usability.
Implemented other miscellaneous updates.
Recompiled executable binaries.

+++ src/DotnetCat/IO/Level.cs
@@ -3,7 +3,7 @@ namespace DotnetCat.IO;
 /// <summary>
 ///  Console output level enumeration type.
 /// </summary>
-internal enum Level : ushort
+internal enum Level : byte
 {
     Error,   // Error/exception messages (stderr)
     Info,    // General messages (stdout)
